requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.0
pillow==10.2.0
requests==2.31.0
openpyxl==3.1.2

sentence-transformers==2.6.1
torch==2.3.0
httpx==0.24.1

